requirements.txt
Flask==1.1.2
openai==0.28.0
python-dotenv==0.19.1

